Move BranchCache and MemStore into their own files
